Update requirements.in to operate on dbt-core 1.7
We wanted to move this project to the latest dbt-core version to ensure
it operates on a version of dbt-core that has addressed the security
issue (CVE-2024-22195) with Jinja2. By association we also had to upgrade
the version of dbt-duckdb being used. Tangentially we also upgraded the
version of sqlfluff.